ser2net 4.6.1: Fixes for separating gensio
Now that gensio is a separate formula, remove all the build of gensio
in ser2net and make it depend on gensio.

Put the ser2net control directory in /opt/homebrew/etc/ser2net instead of
being in the ser2net directory.

Set the datarootdir, where ser2net stores it authentication information,
to /opt/homebrew/share/ser2net.

Run the service in forground as it should be.

Don't enable the admin interface through the command line, the user
can use ser2net.yaml to enable it if they want.

Signed-off-by: Corey Minyard <minyard@acm.org>