automerge-from-merge-queue: fix
The commit that triggers a `merge_group` event is a merge commit that
isn't associated with any PR, so our lookup for the associated PR
doesn't work.

We can fix that by checking the latest PR associated with the second
parent of the merge commit instead. (The first parent is either already
on `master` or is from a merge commit created from another PR in the
merge queue.)