aws-iam-authenticator 0.6.20
This updates `aws-iam-authenticator` to 0.6.20.

The existing `livecheck` block returns 0.5.27 as the newest version
instead of 0.6.20, as the former is marked as the "Latest" release on
GitHub. Since upstream doesn't reliably mark the highest version as
"latest", we have to use the `GithubReleases` strategy instead.

[For what it's worth, we check GitHub releases instead of Git tags
because upstream has labeled a version as "pre-release" in the past.]