bibtex2html: update urls
The existing `homepage` and `stable` URLs for `bibtex2html` currently
redirect from www.lri.fr/~filliatr/ to
usr.lmf.cnrs.fr/~jcf/, so this updates the URLs to avoid the
redirection.

The `livecheck` block is giving an `Unable to get versions` error
because the new server doesn't have an `index.html` file and uses
`index.en.html` instead. This updates the `homepage` URL and brings
the regex up to date (matching the tarball version instead of loose
text on the page).