flyctl: remove throttle
At Fly.io, we've had some trouble with the throttle. Whenever a bug is
released in a version that is divisible by 5, users won't get a bug fix
until the next 5th version. Especially when it comes to critical bug
fixes, it can lead to quite a bit of support headache when users report
bugs that have been fixed in an already released version.

I've talked to my team (who's responsible for flyctl), and we think that
reducing our releases to at most once a day, except for crucial bug
fixes, would be the way to go. We understand that the reason we got
throttled was because having multiple releases a day, and we hope that
releasing less often would be enough to reconsider the throttling :)