tfsec 1.28.9
`tfsec` is on the autobump list but it hasn't been updated to newer
versions because the existing `livecheck` block for `tfsec` checks
the "latest" release on GitHub but the most recent release that has
been marked as "latest" is v1.28.6.

It's not necessary to check the "latest" release here, so this
resolves the issue by updating the `livecheck` block to check Git
tags instead (using the standard regex for tags like `1.2.3`/`v1.2.3`,
to avoid various unwanted tags in the repository).