wireshark: clarify non-GUI in description, add HTTP/3, simplify options
Some users install the wireshark package in CI in order to build binary
Wireshark plugins. The majority of users are likely looking for the
graphical user interface provided by `brew install --cask wireshark`.
Make this clear in the description.

Add libnghttp3 for HTTP/3 header dissection support.

Remove redundant options that are either the default or gone. Homebrew
used to allow optional features, but not anymore.

Disable Logray explicitly. It is currently disabled by default, but
since it pulls in Qt as dependency, let's make that explicit.