openstackclient: add optional pkgs for more features
The OpenStackClient is a unified CLI for a number of OpenStack services
however it does not support all OpenStack services out of the box
without optional dependencies added. This adds a number of them to make
this installation more feature rich for users. It's a resurrection
of #158995 while dropping trove due to the dependency loop. It also
drops clients that don't appear on the current release list at
https://releases.openstack.org/caracal/index.html since the original PR
referenced the list of OpenStack projects from 2016 instead of current
ones.

Signed-off-by: Doug Goldstein <cardoe@cardoe.com>