glooctl 1.17.0
update formula to latest stable glooctl 1.17.0

glooctl changed version output in https://github.com/solo-io/gloo/pull/9212.
This version output requires a different check format for PRs auto-updating
the glooctl formula to pass.

This PR also includes the general update to 1.17.0 stable and an update
to the name of the main branch of the repo.