agda: update livecheck
The existing `livecheck` block for `agda` is giving an unstable
version (2.6.20240714) as the latest version instead of 2.6.4.3. This
happens because the version ending with a date is technically matched
by the standard regex for Git tags like `1.2.3`/`v1.2.3`.

This updates the regex to only match version tags where the last
number is three digits or fewer. This will continue to match stable
version tags while avoiding the tags that end with a date.