jetty: update livecheck
The `livecheck` block for `jetty` was recently updated, as the
previous check wasn't working (due to redirections).

The current regex will only match versions that have a date at the
end, so it will fail to match if new versions don't include a
trailing date. `jetty-runner` appears to have used the same versions
as `jetty` in the past but now it uses versions without a trailing
date.

This updates the regex to make the trailing date optional and to
match it in a less explicit fashion, as there isn't a benefit to
matching a specific number of digits in this context.