jetty-runner: update livecheck
The `livecheck` block for `jetty-runner` was recently updated, as
the previous check wasn't working (due to redirections).

The current regex will only match versions that don't have a date at
the end, so it will fail to match if new versions include a trailing
date. `jetty-runner` appears to have used the same versions as
`jetty` in the past but now it uses versions without a trailing date.

This updates the regex to allow for an optional trailing date, to
align it with the `jetty` regex and ensure that we don't miss
versions due to the version format. It's probably less of a concern
for `jetty-runner` but better safe than sorry.