flvmeta: update stable
The existing `livecheck` block for `flvmeta` checks the homepage but
it no longer contains a tarball link or version information and links
to the GitHub project instead. The existing `stable` tarball still
resolves but if upstream isn't linking to it on the homepage, we
can't identify when a new file is available.

This updates the formula to use a tag tarball from GitHub and removes
the `livecheck` block, so it will use the `Git` strategy by default.