opam 2.2.0-2
By default, livecheck checks the Git tags for `opam` but this returns
2.2.0 as the newest version while the formula uses 2.2.0-1. This is
because upstream has published tarballs with a version suffix as
additional release assets to the existing tag (2.2.0) instead of as
separate releases.

This updates the formula to 2.2.0-2 and adds a `livecheck` block to
check release assets for the "latest" release on GitHub. The regex
will match `opam-full` tarballs with or without a version suffix, so
this will correctly return 2.2.0-2 as the newest version.