fox: add livecheck
livecheck is unable to identify versions for `fox` by default, so this
adds a `livecheck` block that checks the first-party download page.

The page can contain links for development, new stable, stable, and
old stable versions and they all use a stable filename format (e.g.,
fox-1.7.84.tar.gz is development and fox-1.6.58.tar.gz is stable).
Upstream explains in the FAQ that versions with an odd-numbered minor
are development versions and those with an even-numbered minor are
stable (see http://www.fox-toolkit.org/faq.html#VERSION), so I'm using
that in the regex to identify stable versions (instead of trying to
match the trailing `(STABLE)`/`(NEW STABLE)` text, which could be
error-prone).