man2html: add livecheck
By default, livecheck is unable to check `man2html`. This adds a
`livecheck` block that checks the directory listing page where the
`stable` archive is found. Version 3.0.1 appears to be from 1997, so
it seems unlikely that there will be a new version anytime soon (if
ever) but this is technically checkable, so we may as well.