lockrun: add livecheck
livecheck is unable to identify versions for `lockrun` by default, so
this adds a `livecheck` block that matches the version information
from the comments in `lockrun.c`. It's a little silly to download the
`stable` file to identify version information but it's only a few KB,
so it's fine. The newest version (1.1.3) is from 2014-11-03 but this
is technically checkable, so we may as well.