wordnet: add livecheck
livecheck is unable to identify versions for `wordnet` by default.
This adds a `livecheck` block that checks the upstream download page,
which links to both the WordNet tarball and database tarball that the
formula uses. The regex matches versions in both of these filenames,
as these versions may not align (e.g., we're using WordNet 3.0 but
the database version is 3.1, so we use that as the formula version).