nload: update urls, license
The www.roland-riegel.de URLs in the `nload` formula redirect from
HTTP to HTTPS, so this updates the URLs to resolve the redirection.

This also updates the `license` from `GPL-2.0` to `GPL-2.0-or-later`,
as the `README` contains license information using the "or...later"
language (in addition to license comments in the source files).