jxrlib: add livecheck, update version
livecheck is unable to identify versions for `jxrlib` by default. This
adds a `livecheck` block that checks the Debian directory listing page
and massages filename versions to match the formula `version` format.

Besides that, this removes `.orig` from the end of the formula
`version`, as that's a part of Debian's file name format and arguably
not version information.