tdb: update livecheck
A `livecheck` block was recently added to the `tdb` formula that
checks the directory listing page where the `stable` tarball is found.
This updates the regex to match the version from the filename in the
`href` attribute, bringing it in line with typical regex patterns.
[This currently doesn't make much of a functional difference in this
particular case, so it's mostly about standardization.]