moon-buggy: add livecheck
livecheck is unable to identify versions for `moon-buggy` by default.
This adds a `livecheck` block that checks the homepage, which links
to the `stable` tarball. Upstream uses a similar version format for
stable and unstable versions (e.g., 1.0 is stable but 1.0.51 is
experimental), so this identifies stable versions by looking for the
trailing `(stable version)` annotation. This isn't ideal but it works.