kotlin: update livecheck
The existing `livecheck` block for `kotlin` uses the `GithubLatest`
strategy to match the version from the "latest" release on GitHub.
However, this is currently 1.9.25, as upstream is maintaining both
1.x and 2.x versions.

This updates the `livecheck` block to use the `GithubReleases`
strategy, which is necessary to match the highest [stable] version
from recent releases. We can switch back to the `GithubLatest`
strategy if/when upstream stops maintaining multiple major versions
or if the "latest" release reliably points to the highest version
in the future.