janet: improve stability of JPM and modules
The current formula for Janet installs the Janet Project Manager
("JPM"). JPM is Janet's official mechanism for installing Janet modules.
Janet modules are bundles of code that can be either libaries or
executables. Libraries can be source code in Janet that is loaded by a
Janet project directly or source code in other languages that is
compiled to use Janet's C interface. Similarly, executables can be
source code in Janet that is run as executable scripts or source code
in other languages that is compiled to an executable binary on
installation.

Homebrew's formula installs JPM in the keg where Janet itself is
installed. This means that JPM uses the `bin`, `lib` and `share/man`
directories within the keg for modules that _it_ downloads. This is
problematic for three reasons. First, it means that when Janet is
updated, libraries installed previously seem to the user to have have
disappeared. Second, executables are not on the user's PATH after
installation unless the user add's the keg's `bin` directory to their
path (which will in turn break if the respective keg is uninstalled).
Third, man pages installed by JPM do not appear unless the user add's
the keg's `share/man` directory to their manpath.

As with other formulae that install package managers (e.g. Node,
Python), a solution that is more stable across runtime updates is for
the package manager to use one or more directories outside the keg.
This commit does that by configuring JPM during the bootstrap phase to
use the respective directories within `{HOMEBREW_PREFIX}`. The downside
of this approach is that when Janet is uninstalled completely, the files
and directories created by JPM remain. Nevertheless, this approach is
preferred to the status quo.

In addition to improving JPM stability, this commit also removes the
build-time dependency on Meson and Ninja. Janet comes with a
well-designed Makefile that works consistently across macOS and Linux
without the need for other build tools.