xrick: remove
Must be removed due to Homebrew/core license criteria.

From README:
```
I have written the xrick code. However, graphics and maps and sounds
are by the authors of the original Rick Dangerous game, and "Rick
Dangerous" remains a trademark of its owner(s) -- maybe Core Design
(who wrote the game) or FireBird (who published it). As of today, I
have not been successful at contacting Core Design.

This makes it a bit difficult to formally release the whole code,
including data for graphics and maps and sounds, under the terms of
licences such as the GNU General Public Licence. So the code is
released "in the spirit" of the GNU GPL. Whatever that means.
```