vip: add livecheck
livecheck is unable to identify versions for `vip` by default. This
adds a `livecheck` block that checks the `stable` file, which
contains a timestamp near the start of the file (e.g.,
`# @(#)  /u/des/src/duf/vip  1.10  97/11/13  13:45:25`). Checking the
`stable` file feels a bit silly (especially since the most recent
update is from 1997) but it's only a few KB and this is technically
checkable.