libpinyin: add livecheck
By default livecheck checks the Git tags for `libpinyin` and it's
currently returning 2.9.91 as latest but this version is marked as
"pre-release" on GitHub. Upstream appears to use a 90+ patch version
to indicate unstable versions, so this adds a `livecheck` block with
a regex that should only match stable versions.