openssl@3: prevent `brew` from pruning the `certs` and `private` directories
When OpenSSL is unable to verify a TLS certificate using its default CA
bundle, it will fall back to certificates it finds in the `certs`
subdirectory.

If the `certs` subdirectory is missing, it will error out with something
like

    error:80000002:system library::No such file or directory

This obscures the TLS error, which can be confusing for users.[^1] Let's
fix that.

Let's also keep the `private` subdirectory. I'm not certain what it
does, but OpenSSL installs it so we probably dont want `brew cleanup` to
prune it either.

[^1]: See, for example, https://github.com/openssl/openssl/discussions/25172