netdata: update livecheck
The existing `livecheck` block for `netdata` checks the Git tags but
this currently gives 1.99.0 as the latest version, whereas the latest
GitHub release is 1.47.0. It's technically possible to update the
regex to avoid versions with a 90+ patch but the `stable` URL uses a
GitHub release asset, so we can sidestep the tag issue by simply
using the `GithubLatest` strategy.