selenium-server 4.24.0
The `selenium-server` formula is on the autobump list but
`bump-formula-pr` is failing to correctly update the `stable` URL
because the filename version and tag version differ in the existing
URL, so `#update_url` only updates the filename version and not the
tag version. This will continue to fail indefinitely, so it's
necessary to manually bump the version to 4.24.0.