tundra: update livecheck
The existing `livecheck` block for `tundra` uses the `GithubLatest`
strategy but this is currently giving an `Unable to get versions`
error because the latest release is simply named "latest" instead of
a version.

Normally, checking the Git tags should be sufficient for `tundra` but
we have to check GitHub releases for the moment because upstream has
recently tagged some higher versions (v2.23, 2.20, 2.21) that don't
have a corresponding release, so the `Git` strategy can't return the
correct latest release. Since `GithubLatest` doesn't work, we have to
use `GithubReleases` for the time being.