rpcgen: remove as SPDX rejected license
SPDX rejected Sun RPC license due to restriction on redistribution[^1].
Fedora also classified license as not-allowed due to restrictions[^2].

At the same time as removal, we add glibc-based copy in `rpcsvc-proto`,
which ships `rpcgen` after it was relicensed. Alternatively, the OpenBSD
copy[^3] could be distributed, but versioning is more complicated.

If Apple's `rpcgen` is updated to relicensed source, then `rpcgen` may
be re-introduced with an updated `license` stanza.

[^1]: https://github.com/spdx/license-list-XML/issues/906
[^2]: https://gitlab.com/fedora/legal/fedora-license-data/-/blob/main/data/LicenseRef-sun-rpc.toml
[^3]: https://github.com/openbsd/src/commit/cb7760d181b85988239aa31e102cfd38c6ab30ad