protobuf@3: deprecate
This version is no longer supported upstream, so let's deprecate it.

To avoid having to change the `deprecate!` to a `disable!` at some point
in the future, let's just set `disable!` to a future date right now, so
that it is now considered deprectated and will be automatically disabled
in the second half of 2025.