libcoap: Build and install command-line tools.
libcoap-client allows to generate simple requests for retrieval and
modification on a remote COAP server from the commandline, as suggested by the
libcoap upstream. libcoap-server and coap-rd are server applications that could
be useful in application tests.

Also, for optimization, do not build Doxygen documentation, we don't install it
anyway.