spidermonkey@115: deprecate
This will no longer be used as a dependency in any formulae after #190925.

To save us the trouble of switching the `deprecate!` to a `disable!` at
some point, let's just set `disable!` to a future date now so that it
will be deprecated as soon as this PR is merged and then disabled in the
second half of 2025.