wrangler: deprecate
This doesn't seem to build with Erlang 25, 26, or 27 (see #191438).

Upstream also seems to be unresponsive to issues about failed builds, so
let's deprecate it now. To avoid having to change the `deprecate!` to a
`disable!` at some future date, let's just set `disable!` now to a
future date. This will deprecate it today and disable it in the second
half of 2025.