ghostscript: set correct library install names
This ensures that the libraries are built with the install names that we
want.

`brew` will fix up the install names for us even if we don't do this,
but we may as well just build the libraries with the right install names
in the first place so we don't have to fix them up later on.