rebar3: enable compatibility with multiple Erlang versions
`rebar3` is designed to work with multiple versions of Erlang. However,
it can only be used with versions of Erlang that are at least as new as
the one it was built with.

To maximise compatibility with supported versions of Erlang, let's build
with `erlang@25`. We also adjust the `install` method to ensure that
we're always building with the oldest supported version of Erlang.