tofrodos 1.8.3
This updates `tofrodos` to the latest version, 1.8.3, released on
2024-06-14. Before that, the most recent release was 1.7.13 on
2013-10-28.

The homepage now links to a first-party zip and this isn't available
from Debian (only 1.7.13), so this updates the `url` to use the
first-party zip. Since the archive format has changed, this also
updates the `livecheck` block regex to match a tarball or zip.

Lastly, this adds a test that exercises some basic functionality.

Co-authored-by: Carlo Cabrera <30379873+carlocab@users.noreply.github.com>