workflows/triage: cleanup usage of `CI-linux-self-hosted(-deps)?`
Now that we skip recursive dependencies on Linux (except for Linux-only
formulae), most of these are not needed. I've retained the formulae that
are Linux-only (alsa-lib, libva, systemd), and ones with more than 200
direct dependents (the rest).

Not really sure what `souffle` was doing on either list, but we can add
it back if it turns out to be needed.

Finally, I think that this list grew a little out of control, so I've
added @Homebrew/tsc as a CODEOWNER for `triage.yml` to ensure further
additions are given more careful review.