ammonite-repl: update livecheck
The existing `livecheck` block for `ammonite-repl` is returning an
`Unable to get versions` error, as the regex expects a version with an
`-M#` suffix but the current version is 3.0.0. This can be resolved
by making the suffix optional in the regex but the latest release
doesn't have any uploaded assets yet (two days later).

With that in mind, we need to update the `livecheck` block to return
the newest stable version that provides the asset we use in the
formula. We can use the `GithubReleases` strategy but there's a lot
of release data, so the response can be slow, requires more data
transfer, and we may still end up surfacing a new version before it's
considered released.

Instead, this updates the `livecheck` block to check the GitHub asset
URLs on the homepage, as it doesn't appear to be updated until the
release assets are available. Though this is a fairly large page, the
download size is smaller than the GitHub releases API response and
it's static content, so we don't have to worry about slow response
times.