kanata 1.7.0 (new formula)
Kanata is a software key remapper. If you are familiar with mechanical
keyboard's firmware, like QMK, then you can think of kanata as a piece
of software that gives any keyboard such powers.

For example, I use kanata to move my arrow keys to the home row, to get
something similar to vim's jikl movement in _any_ application without
configuration.

Kanata is inspired by and similar to kmonad.

Kanata is implemented in Rust, originally targeted Linux and Windows,
but recently grew MacOS support.

On MacOS, it requires Karabiner VirtualHiDDevice Driver to function.
This dependency is not handled by the current formula at the moment.