hdf5: update livecheck
The existing `livecheck` block for `hdf5` returns 1.14.4 as the
latest version instead of 1.14.5 (or even 1.14.4.3). The version on
the page is 1.14.4-2 but the formula uses a GitHub release asset and
the latest GitHub release is 1.14.5.

This addresses the issue by updating the `livecheck` block to check
GitHub releases. Upstream creates releases for multiple major/minor
versions and they're not always released in sequential order, so we
have to check multiple releases instead of just the "latest" one.