rust: fix broken `rust-lld` in `rustlib`
There is a `rustlib/$TARGET_TRIPLE/bin/rust-lld` that links to an
`@rpath/libLLVM.dylib`. Its only `LC_RPATH` command is
`@loader_path/../lib`.

Trying to run this binary in the current state will not work (because it
won't be able to find `libLLVM`). Let's fix that by adding a symlink so
that it can find the wanted library.