wakeonlan: build an `:all` bottle
The differences between the macOS and Linux bottles are due to being
built by different versions of `Pod::Man`. Let's fix that by building
with the same `perl` everywhere.

This should have no effect on users, because no user will have to build
this from source once it has an `:all` bottle.