argtable3: add livecheck
Upstream uses a tag format including a version and hash (e.g.,
`v3.2.2.f25c624`) and we only use the version part in the formula, so
this adds a `livecheck` block with a regex that omits the hash part
to match the formula version format. This makes it so that livecheck
won't incorrectly surface the tag with the hash as newer than the
same version without the hash.