julia: update livecheck
The existing `livecheck` block for `julia` uses the `GithubLatest`
strategy but this is currently giving 1.10.6 as the newest version
instead of 1.11.1. The lower version corresponds with LTS releases,
so we can't guarantee that the "latest" release on GitHub will be
for the newer stable versions.

This updates the `livecheck` block to check the first-party download
page, as it links to the `stable` tarball on GitHub, so we can check
that without having to resort to using the `GithubReleases` strategy.