pdns: update livecheck
The existing `livecheck` block for `pdns` checks the first-party
download page but this links to 4.9.0 (from 2024-03-15). There have
been subsequent 4.9.1 (2024-05-28) and 4.9.2 (2024-10-01) releases
but the download page wasn't updated for these versions. This updates
the `livecheck` block to check the directory listing page where the
`stable` tarball is found.

This will resolve the aforementioned issue but we should switch back
to checking the download page if/when it's reliably updated with each
version release in the future, as the download page is only ~9 KB
(gzipped) but the directory listing page is ~149 KB (because it
doesn't offer compression) and will increase in size with every
software release (not just for `pdns`).