intercept: add livecheck
By default, livecheck checks the Git tags for `intercept` and this
works at the moment. However, some of the previous releases use a
stable version format (e.g., `v1.0.9`) but are marked as "pre-release"
on GitHub. This adds a `livecheck` block that uses the `GithubLatest`
strategy, so livecheck won't surface any pre-release versions.