libdom: add livecheck
livecheck is unable to identify versions for `libdom` by default, so
this adds a `livecheck` block that checks the directory listing page
where the `stable` tarball is found. For what it's worth, there
doesn't appear to be a first-party download page for `libdom` and
this is the closest thing to it.