boring 0.7.0
This commit updates the boring formula to version 0.7.0. It additionally introduces
a post_install hook which kills an existing boring daemon instance (if running). This
ensures that CLI and daemon have the same version. Upon first invocation of the updated
binary, it automatically spawns a new daemon (as in previous versions).