knot: update livecheck, homepage urls
The existing `livecheck` URL for `knot` redirects from secure.nic.cz
to a new www.knot-dns.cz URL. This resolves the redirection by
updating the URL to check the first-party download page instead of
the directory listing page where the `stable` archive is found.

This ensures that livecheck won't surface a new version until upstream
publishes it (not just when they upload the tarball). The directory
listing page also requires more data transfer and will naturally
continue to grow over time as more versions are added, so it's not our
first choice when we have better options.

This also updates the `stable` URL to resolve a similar redirection.