ospray: remove linux bottle
We used to hack `ispc` in order to get it to build on Linux, but hacking
it further was made very difficult by the version update to 1.25.2 to we
lost the Linux bottle in #196480.

This currently has a runtime dependency on `ispc`, so we need to remove
the Linux bottle to prevent dependent testing failures on Linux (e.g.
when `tbb` is updated).