git-spice: update license
When git-spice was addee to homebrew-core (#191536),
it was being distributed under a dual license:
the bulk of it was GPL-3.0-or-later,
except two files which were under BSD-3-Clause.

Those files have since been moved to a separate repository
and deleted from the main repository.
git-spice is now distributed entirely under GPL-3.0-or-later.
(https://github.com/abhinav/git-spice/commit/74b194bf0bedabeddaaf1029942d5139fba06a33)

This updates the license in the formula to match.