wasmtime 26.0.1
This updates `wasmtime` to the latest version, 26.0.1.

This also updates the `livecheck` block to use the `GithubReleases`
strategy, as upstream maintains multiple major versions and the
release for 25.0.3 is marked as "latest" instead of the later 26.0.1
release. Upstream seems to be careful about creating the releases in
order from lowest major version to highest but there are periods where
the newest release is for a lower major version, which can cause
problems for the `GithubLatest` strategy. `GithubReleases` requires
more data transfer and is notably slower (unless cached) but the
upstream website doesn't provide any version information (it simply
links to the GitHub releases page), so this appears to be our only
option unless/until this situation changes.