tsduck 3.39-3956 (updated formula)
Added dependency to "bash" on macOS to enforce the use of bash v5 instead
of the old macOS-provided bash. See discussion in PR #197908 which
explains why this is necessary.

No need to regenerate the bottles. A fix was provided to pass the build
and already released version 3.39-3956. This PR fixes the build problem
in Homebrew for the next version but does not change the content of the
project.