workflows/autobump: adjust cron schedule
The autobump workflow run for core can overlap with that for cask, which
makes it easier to hit API rate limit. Let's adjust the cron schedule to
reduce the chance of that happening.

For reference, the autobump workflow for cask is scheduled for
`23 */3 * * *` (every 3 hours, with an offset of 23 minutes).