pdal: update livecheck
The existing `livecheck` block for `pdal` checks the first-party
download page but it hasn't been updated to link to the 2.8.1 release
on GitHub, so the formula version is newer than the livecheck version
(2.8.0). This updates the `livecheck` block to use the `GithubLatest`
strategy, as the `stable` tarball is a GitHub release asset.

Regarding the `livecheck` block comment, upstream stopped creating
tags that only use a major/minor version after 2.3 (2021-05-19). From
that release onward upstream has consistently used a major/minor/patch
version for both the release tag and filename, so we don't have to
worry about that now.