php: Remove unknown configure flags and add SNMP
The pspell extension has been removed in 8.4. See https://www.php.net/manual/en/migration84.removed-extensions.php
This removes the dependency on aspell and the configure flag `--with-pspell`.

The configure flag `--with-kerberos` is also unrecognized (warning in configure output).

Add the SNMP extension back and add test. It was only broken back when we had mixed crypto linking in old macOS.