libsidplayfp 2.12.0 (new formula)
Adds formula for libsidplayfp, an active fork of sidplay2. It is
an emulator for the Commodore 64 chipset to play C64 SID music.

libsidplayfp is packaged by all major Linux distribution (Debian,
Fedora, Arch, and derivatives). It is an optional dependency
of the audacious music player available in homebrew.

libsidplayfp: update test

Co-authored-by: Rui Chen <rui@chenrui.dev>