bilix: remove livecheck
The current `bilix` `livecheck` block was recently added because
livecheck was using the `Pypi` strategy by default but it was failing
because of upstream changes. We've made changes to the `Pypi` strategy
to work around the issue, so this removes the `livecheck` block to
ensure that livecheck will use the `Pypi` for this formula again.