rpds-py 0.22.3 (new formula)
For GNU Radio, we need jsonschema, which in turn depends on rpds-py,
which in turn depends on Rust. It seemed desirable to move that
dependency (since it's generally useful and doesn't need to be rebuilt
for every GNU Radio build) into its own formula. See GH #201049.

This supersedes #178716. Unlike that, it uses the existing maturin
formula instead of doing a pypi-based source build.

Signed-off-by: Marcus Müller <mmueller@gnuradio.org>