gnupg: update livecheck
livecheck is currently returning 2.5.2 as the latest version for
`gnupg` but this is an unstable version according to the first-party
download page (https://gnupg.org/download/). Looking at the
"End-of-life Announcements" section of that page, upstream seems to
indicate stable versions using an even-numbered minor version. With
that in mind, this updates the `livecheck` block regex to restrict
matching to tarball versions with an even-numbered minor.