forgit: add livecheck
By default, livecheck checks the Git tags for `forgit` and this works
but the formula's `stable` URL is a GitHub release asset and there
can be a notable gap between when a tag is created and a release
appears.

For example, 24.12.0 was tagged on 2024-11-30 but there hasn't been a
release yet, so livecheck has been returning this version as newest
even though we can't update to it until a GitHub release is available.