duck: update livecheck
The `livecheck` block for `duck` is currently returning an `Unable to
get versions` error. The `duck-src` tarball links on the dist.duck.sh
directory listing page have disappeared but the URLs still resolve.
It's unclear whether a `duck-src` URL will be available for the next
version but this loosens the `livecheck` block regex to make the
`-src` text optional, so this will surface versions from the `duck`
tarballs. This fixes the check, so we will at least know when a new
version is available and can make any necessary changes then.