sstp-client: update homepage
The SourceForge project for `sstp-client` states that the project has
moved to https://gitlab.com/eivnaes/sstp-client. The existing
`homepage` is a SourceForge site and it links to the SourceForge
files as the download location, so it's not up to date. The `stable`
and `head` URLs have already been updated, so this updates the
`homepage` to use the GitLab project.