bluez: add livecheck, head
By default, livecheck isn't able to check the `stable` URL for
`bluez`, so it falls back to checking the Git tags from the `homepage`
URL. This adds a `livecheck` block that checks the directory listing
where the `stable` tarball is found. Besides that, this updates the
`homepage` to bluez.org.

While we're at it, this adds a `head` URL and makes sure the `HEAD`
build works.