git-if: add livecheck
By default, livecheck isn't able to check the `stable` URL for
`git-if`, so it falls back to checking the Git tags from the `head`
URL. This adds a `livecheck` block that checks the homepage, which
links to the `stable` archive. The archive filename uses a dotless
version, so this matches the version from the "Git 1.2.3" text after
the archive link.

This also updates the `homepage` link, as it redirects from the
existing URL to this new one.