glulxe: add livecheck
By default, livecheck isn't able to check the `stable` URL for
`glulxe`, so it falls back to checking the Git tags from the `head`
URL. This adds a `livecheck` block that checks the homepage, which
links to the `stable` tarball. The tarball filename uses a dotless
version, so this matches the version from the "Glulxe 1.2.3" inner
text of the link.