gnu-getopt: add livecheck
By default, livecheck isn't able to check the `stable` URL for
`gnu-getopt`, so it falls back to checking the Git tags from the
`head` URL. This adds a `livecheck` block that checks the directory
listing page where the `stable` tarball is found. Unfortunately the
version directories only include the major/minor version and it's
necessary to check within the directory to identify the full version
from the tarball links, so this requires two requests.